fix(vimrc): don't conceal json/markdown
I write a blog, the contents of which are written purely in markdown and
for the life of me I could not figure out why suddenly some of the
contents of my file ...disappeared... Then I found out that the values
are still there but they are hidden.
For the time being, I just disabled the syntax highlighting and
continued working. (:syntax off)
The actual issue was related to conceal settings.
The default conceal level was set to 2 due indent lines + polyglot
combination. To resolve this I added these changes to my config.