fix(experimental): substitute in visual selection
I was unable to understand why sometimes the <leader>x map would work
like normal x in visual mode (i.e. delete the selection). Then I looked
at the mapping again, and it was clear that since I used 'n'noremap, I
was actually only allowed to use this map in normal mode.
To fix this, first I changed this to noremap (effectively use this map
in all modes), then it was still not working as expected. I was getting
an error regarding trailing characters. The reason was to substitute in
a visual block I first have to switch to normal mode. So, I added an
<Esc> before the command, and now it works as expected.
Next I want to set a different colorscheme for visual mode, because it
doesn't look friendly to my eyes currently.

Reference:
https://vim.fandom.com/wiki/Search_and_replace_in_a_visual_selection