feat(config): show <current/total> match count
As it turns out, the vim-cool plugin has an <experimental> feature which
enables a counter for search.
I was looking at someone's config, then stumbled upon a plugin that does
this. The plugin itself referenced the vim-cool plugin as one of the
alternatives, and that's how I ended up adding this to my config.