feat(config): update auto-pairs plugin
I needed a feature to blacklist my vim config folder from
the AutoPairs plugin. @LunarWatcher saw the issue that I
opened and added the feature in his fork of the Plugin.
Hence, I switched to it. Added the config to use the alt-e
mapping. added config to disable plugin in vim-config folder.

closes jiangmiao/auto-pairs#310