feat(treesitter): better syn-hl with treesitter
Treesitter is a parser which uses grammar for different languages to
parse the current buffer and generates the syntax tree. It is fast,
efficient and supported natively in nvim. One of the main advantages
(for me personally) is better syntax highlighting in Python files.

I experimented with a couple of things after looking at a video by
@ChristianChiarulli.
I ended up using one of the colorschemes in his plugin itself
I am trying out this: `snazzy` colorscheme for now.
It definitely has better highlighting but I feel that there is some
confusion regarding docstrings. For now, I am ok with how the scheme
looks though.
Point to note here is that only snazzy is working properly for Python
(highlighting self, def, if, variable names etc.). All the others are
missing something or the other.

Note: upgraded neovim to 0.5x
- this is causing problems with undo dir
- there are probably breaking changes due to this, meaning vim & nvim
  can no longer support the same undo format. Need to figure out if I
  want to stick with nvim or not.