fix(config): fzf preview.sh for windows
The original preview.sh file bundled with fzf was causing problems with
previewing in windows.
I no longer remember the change which fixes the issue, but I think I
removed blank lines, and installed `bat` in WSL to get it work the way I
wanted it to.
Adding this file here as a reference in case I face similar problem in
future on a different machine.

References:
https://github.com/junegunn/fzf.vim/pull/1189
Closes: https://github.com/junegunn/fzf.vim/issues/1207