feat(theme): default scheme set to solarized
this is an experiment, but has been going on for a couple days now. I
really like the vibe of the screen with this scheme. It's very different
from a completely 'black' background. Although I haven't had anyone say
they like (or dislike) the theme. Come on people, how are you not seeing
this drastic change in my config? (we work with screenshare on from time
to time)