feat(config): fzf should look for hidden files too
with my original config, fzf would always ignore files in my .gitignore
However, recent projects made it so that some files are to be kept in
local mahcines, but can't be commited. I still want work on them, look
them up, change them on the local system, so these files need to be
searchable.
Since I use ripgrep + fzf combination, all I needed to do was add in the
--no-ignore-vcs flag (see patch for this commit)

Reference:
https://github.com/junegunn/fzf.vim/issues/121#issuecomment-546360911