feat(plugin): firenvim specific config
firenvim opens an instance of neovim in the browser where the input
field is a textarea.
Although it is nice to have the full functionality of neovim, there are
some settings that need to be changed. For example, we don't want to see
the statusline, and line numbers because the text area is already very
small.
Also, more often than not, I am writing markdown in text-areas, so
setting the filetype to markdown helps (although the jury still out on
this one, maybe that setting should just be limited to github.com)